+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 1.0.0-beta
+
+**Added:**
+
+- Now you can specify any element as a "trigger element" for the picker. While doing that you have to specify an input (type text) element in `options` by "`input`" key that would hold the value.
+- You can hit `SpaceBar` or `Enter` key on Meridiem (AM/PM) buttons to confirm.
+
+**Documentation:**
+
+- LICENCE.md file updated.
+- README.MD file updated.
+
+**Fixed:**
+
+- `input` event conflicted with `keydown` event, which prevented typing/entering numbers in the picker.
+
 ## 1.0.0-alpha.0
 
 - Initial alpha release.